+++ app/src/main/java/com/hellohasan/bongabdosample/screen/home/BongabdoCalculation.kt
@@ -3,9 +3,9 @@ package com.hellohasan.bongabdosample.screen.home
 import android.icu.util.Calendar
 import com.hellohasan.bongabdo.api.Bongabdo
 import com.hellohasan.bongabdo.api.BongabdoMethod
-import com.hellohasan.bongabdo.localization_config.BengaliLocalizationConfig
-import com.hellohasan.bongabdo.localization_config.BongabdoLocalizationConfig
-import com.hellohasan.bongabdo.localization_config.EnglishLocalizationConfig
+import com.hellohasan.bongabdo.localization.BengaliLocalizationConfig
+import com.hellohasan.bongabdo.localization.BongabdoLocalizationConfig
+import com.hellohasan.bongabdo.localization.EnglishLocalizationConfig
 import com.hellohasan.bongabdosample.screen.settings.LocaleManager
 import java.text.SimpleDateFormat
 import java.util.Locale
@@ -19,7 +19,7 @@ object BongabdoCalculation {
      */
     fun getBanglaAcademyBongabdo(calendar: Calendar): String {
         val bongabdo = Bongabdo.getInstance(BongabdoMethod.BANGLA_ACADEMY)
-        bongabdo.mLocalizationConfig = getLocalizationConfig()
+        bongabdo.localizationConfig = getLocalizationConfig()
         val bongabdoDate = bongabdo.getBongabdoData(
             calendar.get(Calendar.YEAR),
             calendar.get(Calendar.MONTH),
@@ -30,7 +30,7 @@ object BongabdoCalculation {
 
     fun getDrikShiddhantaBongabdo(calendar: Calendar): String {
         val bongabdo = Bongabdo.getInstance(BongabdoMethod.INDIAN_DRIK_SIDDHANTA)
-        bongabdo.mLocalizationConfig = getLocalizationConfig()
+        bongabdo.localizationConfig = getLocalizationConfig()
         val bongabdoDate = bongabdo.getBongabdoData(
             calendar.get(Calendar.YEAR),
             calendar.get(Calendar.MONTH),
