+++ bongabdo/build.gradle.kts
@@ -23,7 +23,7 @@ publishing {
         create<MavenPublication>("maven") {
             groupId = "com.github.hasancse91"
             artifactId = "bongabdo"
-            version = "1.0.1"
+            version = "1.0.2"
             from(components["kotlin"])
         }
     }
