+++ gradle/libs.versions.toml
@@ -2,14 +2,14 @@
 agp = "8.7.1"
 kotlin = "2.0.21"
 coreKtx = "1.13.1"
-junitJupiter = "5.8.2"
+junitJupiter = "5.10.3"
 junitVersion = "1.2.1"
 espressoCore = "3.6.1"
 kotlinxDatetime = "0.6.1"
 lifecycleRuntimeKtx = "2.8.6"
 activityCompose = "1.9.3"
 composeBom = "2024.10.00"
-jetbrainsKotlinJvm = "1.9.0"
+jetbrainsKotlinJvm = "1.9.25"
 materialIconsExtended = "1.7.4"
 navigationCompose = "2.8.3"
 appcompat = "1.7.0"
